Fix(cmd/bump): Fix length caluclation error.
If you had three items in the queue, `sq bump` would attempt to generate a
random number between 1 (the minimum that isn't the head of the queue) and 2.
Due to the way Rust's ranges are implemented, 1..2 is an empty range as the
range syntax is exclusive. Changing the syntax to be 1..=2 makes it work as
intuition would expect.