Fix: Save last calculation date.
This is an interesting one, because it caused the timestamp of the last
calculation to not be written, at any point, causing SQ to always do the
calculation correctly (from the specified timestmap), but never writing the
correct current value back, leading to purchasing bugs. The amount the purchase
command tried to reckon with was the one from the last write, which was pre
queues update. This caused it tell me that I had $200, but couldn't buy a $45
item. This fixes the caluclation, making the system work as intended again.